Move DevToolsExtensions to the extensions package
